removed obsolete compacting workaround
rollup-plugin-cleanup v2.0.1 includes a fix for the bug that we worked
around in 7d064b12d89b1a67f11fa61424f20136544545b9.

This change declares that we require the version with that fix, removes
the configuration working around the bug, and adapts the tests so they
expect the plugin to actually run.